refactor: video layout move to custom
