+++ components/custom/index.ts
@@ -2,4 +2,6 @@ export { default as CustomInput } from "./input";
 export { default as CustomCheckBox } from "./radio-box";
 export { default as CustomNicknameEditForm } from "./nickname-edit-form";
 export { default as CustomSelectBox } from "./select-box";
-export { default as AlertModal } from "./alert-modal"
+export { default as AlertModal } from "./alert-modal";
+
+export { default as VideoList } from "./video-list";
