+++ components/home/index.ts
@@ -1,2 +1,1 @@
 export { default as HomeCategory } from "./category";
-export { default as HomeVideos } from "./videos";
