+++ pages/index.tsx
@@ -1,12 +1,13 @@
 import { NextPageWithLayout } from "types/common";
-import { HomeCategory, HomeVideos } from "components/home";
+import { HomeCategory } from "components/home";
+import { VideoList } from "components/custom";
 
 const Home: NextPageWithLayout = () => {
   return (
     <div className="layout-p">
       <h3 id="sub-title">주간 인기 동영상</h3>
       <HomeCategory />
-      <HomeVideos />
+      <VideoList />
     </div>
   );
 };
